Refactor Theme Studio components and add installation status handling in macOS app
