+++ apps/control-center/src/components/theme-studio-screen.tsx
@@ -15,13 +15,8 @@ import {
   RefreshCw,
   Send,
   Square,
-  Trash2,
   Type,
 } from "lucide-react";
-import type {
-  PointerEvent as ReactPointerEvent,
-  ReactNode,
-} from "react";
 import {
   useCallback,
   useEffect,
@@ -33,7 +28,6 @@ import {
 import {
   buildThemePack,
   createStarterThemeSpec,
-  deviceThemeSpecJson,
   importThemeSpec,
   normalizeThemeSpec,
   referencedThemeAssetPaths,
@@ -46,12 +40,9 @@ import {
 import {
   assetFileName,
   assetKind,
-  assetKindLabel,
   fileToBase64,
-  formatBytes,
   importSpriteFile,
   spriteMetadata,
-  themeAssetByteLength,
   themeAssetPathForFile,
 } from "@/lib/theme-studio-assets";
 import {
@@ -73,12 +64,7 @@ import {
   AdvancedPanel,
   type ThemeStudioAdvancedTab,
 } from "./theme-studio/advanced-panel";
-import {
-  ColorField,
-  NumberField,
-  SelectField,
-  TextField,
-} from "./theme-studio/editor-fields";
+import { ColorField, TextField } from "./theme-studio/editor-fields";
 import {
   StatusLine,
   StatusPill,
@@ -86,97 +72,36 @@ import {
 } from "./theme-studio/editor-status";
 import { LeaveEditorDialog } from "./theme-studio/leave-editor-dialog";
 import { ThemeStudioToolbar } from "./theme-studio/theme-studio-toolbar";
+import { EditableThemePreview } from "./theme-studio/editable-theme-preview";
+import { PrimitiveInspector } from "./theme-studio/primitive-inspector";
 import {
-  ThemeSpecPreview,
-  type ThemeRenderPack,
-} from "./live-vibetv-preview";
-import { ControlCenterButton } from "./control-center-button";
+  AddButton,
+  AssetRow,
+  DarkButton,
+  PanelTitle,
+} from "./theme-studio/editor-controls";
+import {
+  clampInt,
+  clampedMoveDelta,
+  defaultPrimitive,
+  DISPLAY_SIZE,
+  isEditableKeyboardTarget,
+  normalizeSelectedIndices,
+  prettyJson,
+  primitiveBounds,
+  primitiveTitle,
+  setPrimitiveField,
+  textPrimitiveFontSizeFromVisualHeight,
+  textPrimitiveNaturalWidth,
+  titleFromThemeId,
+} from "./theme-studio/editor-geometry";
+import type { ThemeRenderPack } from "./live-vibetv-preview";
 import { themeRenderPackUrl } from "./control-center-runtime";
 
-type DragState =
-  | {
-      mode: "move";
-      origins: DragMoveOrigin[];
-      startX: number;
-      startY: number;
-    }
-  | {
-      currentX: number;
-      currentY: number;
-      mode: "select";
-      startX: number;
-      startY: number;
-    }
-  | {
-      edgeOffsetX: number;
-      edgeOffsetY: number;
-      index: number;
-      mode: "resize";
-      originHeight: number;
-      originWidth: number;
-      originX: number;
-      originY: number;
-    };
-
-type ResizeSize = {
-  height: number;
-  width: number;
-};
-
-type DragMoveOrigin = {
-  height: number;
-  index: number;
-  width: number;
-  x: number;
-  y: number;
-};
-
-type PrimitiveMove = {
-  index: number;
-  x: number;
-  y: number;
-};
-
-type SelectionBox = {
-  height: number;
-  width: number;
-  x: number;
-  y: number;
-};
-
-type FieldKey = keyof ThemeStudioPrimitive;
-
-const DISPLAY_SIZE = 240;
 const COLOR_FALLBACK = "#000000";
 const DEFAULT_GIF_SIZE = 80;
-const DEFAULT_SPRITE_FPS = 8;
 const MAX_TEXT_FONT_SIZE = 30;
-const TEXT_SELECTION_HEIGHT_SCALE = 1.2;
-const DEFAULT_FRAME = {
-  v: 2,
-  provider: "vibetv",
-  label: "VibeTV",
-  session: 62,
-  weekly: 62,
-  resetSecs: 3600,
-  usageMode: "remaining",
-  activity: "preview",
-  sessionTokens: 12000,
-  weekTokens: 82000,
-  totalTokens: 248000,
-  time: "12:00",
-  date: "03.07",
-};
 const RETIRED_AI_THEME_STORAGE_PREFIX = "vibetv.controlCenter.aiTheme";
-
-const VARIABLE_TOKENS = [
-  { label: "Label", token: "{label}" },
-  { label: "Session", token: "{session}" },
-  { label: "Weekly", token: "{weekly}" },
-  { label: "Reset", token: "{reset}" },
-  { label: "Mode", token: "{usageMode}" },
-  { label: "Time", token: "{time}" },
-];
 
 export type ThemeStudioEditorSource = "blank" | "custom" | "published";
 
@@ -1560,696 +1485,6 @@ export function ThemeStudioScreen({
   );
 }
 
-function EditableThemePreview({
-  onInteractionCancel,
-  onInteractionCommit,
-  onInteractionStart,
-  onMoveMany,
-  onResize,
-  onSelect,
-  onSelectMany,
-  pack,
-  selectedIndex,
-  selectedIndices,
-  spec,
-}: {
-  onInteractionCancel: () => void;
-  onInteractionCommit: () => void;
-  onInteractionStart: () => void;
-  onMoveMany: (moves: PrimitiveMove[]) => void;
-  onResize: (index: number, size: ResizeSize) => void;
-  onSelect: (index: number, additive?: boolean) => void;
-  onSelectMany: (indices: number[]) => void;
-  pack: ThemeRenderPack;
-  selectedIndex: number;
-  selectedIndices: number[];
-  spec: ThemeStudioSpec;
-}) {
-  const svgRef = useRef<SVGSVGElement>(null);
-  const dragRef = useRef<DragState | null>(null);
-  const [selectionBox, setSelectionBox] = useState<SelectionBox | null>(null);
-  const [prefersReducedMotion, setPrefersReducedMotion] = useState(() =>
-    typeof window === "undefined"
-      ? false
-      : window.matchMedia("(prefers-reduced-motion: reduce)").matches,
-  );
-
-  useEffect(() => {
-    const media = window.matchMedia("(prefers-reduced-motion: reduce)");
-    const handleChange = (event: MediaQueryListEvent) =>
-      setPrefersReducedMotion(event.matches);
-    media.addEventListener("change", handleChange);
-    return () => media.removeEventListener("change", handleChange);
-  }, []);
-
-  function pointerPoint(event: ReactPointerEvent<SVGElement>) {
-    const rect = svgRef.current?.getBoundingClientRect();
-    if (!rect) {
-      return { x: 0, y: 0 };
-    }
-    return {
-      x: ((event.clientX - rect.left) / rect.width) * DISPLAY_SIZE,
-      y: ((event.clientY - rect.top) / rect.height) * DISPLAY_SIZE,
-    };
-  }
-
-  function startDrag(
-    event: ReactPointerEvent<SVGRectElement>,
-    index: number,
-  ) {
-    event.preventDefault();
-    event.stopPropagation();
-    const point = pointerPoint(event);
-    const primitive = spec.primitives[index];
-    if (isFullCanvasRect(primitive)) {
-      startSelection(event);
-      return;
-    }
-    const normalizedSelection = normalizeSelectedIndices(
-      selectedIndices,
-      spec.primitives.length,
-    );
-    const shouldMoveSelection =
-      normalizedSelection.length > 1 &&
-      normalizedSelection.includes(index) &&
-      !event.shiftKey &&
-      !event.metaKey &&
-      !event.ctrlKey;
-    const moveIndices = shouldMoveSelection ? normalizedSelection : [index];
-    dragRef.current = {
-      mode: "move",
-      origins: moveIndices.flatMap((moveIndex) => {
-        const movePrimitive = spec.primitives[moveIndex];
-        if (!movePrimitive) {
-          return [];
-        }
-        const bounds = primitiveBounds(movePrimitive);
-        return [{
-          height: bounds.height,
-          index: moveIndex,
-          width: bounds.width,
-          x: movePrimitive.x,
-          y: movePrimitive.y,
-        }];
-      }),
-      startX: point.x,
-      startY: point.y,
-    };
-    onInteractionStart();
-    if (!shouldMoveSelection) {
-      onSelect(index, event.shiftKey || event.metaKey || event.ctrlKey);
-    }
-    event.currentTarget.setPointerCapture(event.pointerId);
-  }
-
-  function startSelection(event: ReactPointerEvent<SVGRectElement>) {
-    event.preventDefault();
-    const point = pointerPoint(event);
-    dragRef.current = {
-      currentX: point.x,
-      currentY: point.y,
-      mode: "select",
-      startX: point.x,
-      startY: point.y,
-    };
-    setSelectionBox(normalizedSelectionBox(point.x, point.y, point.x, point.y));
-    event.currentTarget.setPointerCapture(event.pointerId);
-  }
-
-  function startResize(
-    event: ReactPointerEvent<SVGElement>,
-    index: number,
-  ) {
-    event.preventDefault();
-    event.stopPropagation();
-    const point = pointerPoint(event);
-    const primitive = spec.primitives[index];
-    const bounds = primitiveBounds(primitive);
-    dragRef.current = {
-      edgeOffsetX: point.x - (primitive.x + bounds.width),
-      edgeOffsetY: point.y - (primitive.y + bounds.height),
-      index,
-      mode: "resize",
-      originHeight: bounds.height,
-      originWidth: bounds.width,
-      originX: primitive.x,
-      originY: primitive.y,
-    };
-    onInteractionStart();
-    onSelect(index, event.metaKey || event.ctrlKey);
-    event.currentTarget.setPointerCapture(event.pointerId);
-  }
-
-  function movePointer(event: ReactPointerEvent<SVGElement>) {
-    const drag = dragRef.current;
-    if (!drag) {
-      return;
-    }
-    event.preventDefault();
-    const point = pointerPoint(event);
-    if (drag.mode === "select") {
-      drag.currentX = point.x;
-      drag.currentY = point.y;
-      setSelectionBox(
-        normalizedSelectionBox(drag.startX, drag.startY, point.x, point.y),
-      );
-      return;
-    }
-    if (drag.mode === "resize") {
-      const primitive = spec.primitives[drag.index];
-      if (!primitive) {
-        return;
-      }
-      const maxWidth = DISPLAY_SIZE - drag.originX;
-      const maxHeight = DISPLAY_SIZE - drag.originY;
-      const freeSize = {
-        height: clampInt(point.y - drag.originY - drag.edgeOffsetY, 1, maxHeight),
-        width: clampInt(point.x - drag.originX - drag.edgeOffsetX, 1, maxWidth),
-      };
-      onResize(
-        drag.index,
-        event.shiftKey
-          ? aspectLockedResizeSize({
-              maxHeight,
-              maxWidth,
-              originHeight: drag.originHeight,
-              originWidth: drag.originWidth,
-              targetHeight: freeSize.height,
-              targetWidth: freeSize.width,
-            })
-          : freeSize,
-      );
-      return;
-    }
-    if (drag.origins.length === 0) {
-      return;
-    }
-    const delta = clampedMoveDelta(
-      drag.origins,
-      point.x - drag.startX,
-      point.y - drag.startY,
-    );
-    onMoveMany(
-      drag.origins.map((origin) => ({
-        index: origin.index,
-        x: origin.x + delta.x,
-        y: origin.y + delta.y,
-      })),
-    );
-  }
-
-  function stopDrag(outcome: "cancel" | "commit" | "selection" = "commit") {
-    const drag = dragRef.current;
-    dragRef.current = null;
-    setSelectionBox(null);
-    if (!drag || drag.mode === "select" || outcome === "selection") {
-      return;
-    }
-    if (outcome === "cancel") {
-      onInteractionCancel();
-    } else {
-      onInteractionCommit();
-    }
-  }
-
-  function finishPointer(event: ReactPointerEvent<SVGElement>) {
-    const drag = dragRef.current;
-    if (!drag) {
-      return;
-    }
-    if (drag.mode === "select") {
-      event.preventDefault();
-      const box = normalizedSelectionBox(
-        drag.startX,
-        drag.startY,
-        drag.currentX,
-        drag.currentY,
-      );
-      onSelectMany(
-        box.width < 2 && box.height < 2
-          ? []
-          : selectedPrimitiveIndices(spec.primitives, box),
-      );
-    }
-    stopDrag(drag.mode === "select" ? "selection" : "commit");
-  }
-
-  useEffect(() => {
-    function handleEscape(event: KeyboardEvent) {
-      if (event.key !== "Escape" || !dragRef.current) {
-        return;
-      }
-      event.preventDefault();
-      stopDrag("cancel");
-    }
-    window.addEventListener("keydown", handleEscape);
-    return () => window.removeEventListener("keydown", handleEscape);
-  });
-
-  return (
-    <div className="relative aspect-square w-full max-w-[480px] overflow-hidden border border-[#1B1B1B] bg-black p-0">
-      <ThemeSpecPreview
-        animate={!prefersReducedMotion}
-        pack={pack}
-        status="ready"
-        themeId={spec.themeId}
-      />
-      <svg
-        aria-label="Editable 240x240 preview"
-        className="absolute inset-0 h-full w-full [touch-action:none]"
-        onPointerCancel={() => stopDrag("cancel")}
-        onPointerMove={movePointer}
-        onPointerUp={finishPointer}
-        ref={svgRef}
-        viewBox="0 0 240 240"
-      >
-        <rect
-          aria-hidden="true"
-          className="cursor-crosshair"
-          fill="transparent"
-          height={DISPLAY_SIZE}
-          onPointerDown={startSelection}
-          width={DISPLAY_SIZE}
-          x="0"
-          y="0"
-        />
-        {spec.primitives.map((primitive, index) => {
-          const bounds = primitiveBounds(primitive);
-          const selected = selectedIndices.includes(index);
-          const active = selectedIndex === index;
-          return (
-            <g key={`${primitive.type}-${index}`}>
-              <rect
-                aria-label={`Select ${primitive.type} ${index + 1}`}
-                className="cursor-move"
-                fill="transparent"
-                height={Math.max(8, bounds.height)}
-                onKeyDown={(event) => {
-                  if (event.key !== "Enter" && event.key !== " ") {
-                    return;
-                  }
-                  event.preventDefault();
-                  onSelect(
-                    index,
-                    event.shiftKey || event.metaKey || event.ctrlKey,
-                  );
-                }}
-                onPointerDown={(event) => startDrag(event, index)}
-                role="button"
-                stroke={selected ? "#C7FF68" : "transparent"}
-                strokeDasharray={active ? "4 3" : "2 2"}
-                strokeWidth={selected ? (active ? 1.5 : 1) : 0}
-                tabIndex={0}
-                width={Math.max(8, bounds.width)}
-                x={primitive.x}
-                y={primitive.y}
-              />
-              {active ? (
-                <g
-                  aria-hidden="true"
-                  className="cursor-se-resize"
-                  onPointerDown={(event) => startResize(event, index)}
-                >
-                  <rect
-                    fill="transparent"
-                    height="14"
-                    pointerEvents="all"
-                    width="14"
-                    x={primitive.x + Math.max(8, bounds.width) - 7}
-                    y={primitive.y + Math.max(8, bounds.height) - 7}
-                  />
-                  <circle
-                    cx={primitive.x + Math.max(8, bounds.width)}
-                    cy={primitive.y + Math.max(8, bounds.height)}
-                    fill="#C7FF68"
-                    pointerEvents="none"
-                    r="3"
-                  />
-                </g>
-              ) : null}
-            </g>
-          );
-        })}
-        {selectionBox ? (
-          <rect
-            fill="#CCFF0033"
-            height={selectionBox.height}
-            pointerEvents="none"
-            stroke="#CCFF00"
-            strokeDasharray="4 3"
-            strokeWidth="1"
-            width={selectionBox.width}
-            x={selectionBox.x}
-            y={selectionBox.y}
-          />
-        ) : null}
-      </svg>
-    </div>
-  );
-}
-
-function PrimitiveInspector({
-  onChange,
-  onDelete,
-  onInsertToken,
-  primitive,
-}: {
-  onChange: (field: FieldKey, value: unknown) => void;
-  onDelete: () => void;
-  onInsertToken: (token: string) => void;
-  primitive: ThemeStudioPrimitive;
-}) {
-  const bounds = primitiveBounds(primitive);
-  return (
-    <div className="grid gap-3">
-      <div className="grid grid-cols-2 gap-2">
-        <NumberField
-          label="X"
-          value={primitive.x}
-          onChange={(value) => onChange("x", value)}
-        />
-        <NumberField
-          label="Y"
-          value={primitive.y}
-          onChange={(value) => onChange("y", value)}
-        />
-      </div>
-
-      {(primitive.type === "rect" ||
-        primitive.type === "progress" ||
-        primitive.type === "gif" ||
-        primitive.type === "sprite" ||
-        primitive.type === "pixels" ||
-        primitive.width !== undefined) ? (
-        <div className="grid grid-cols-2 gap-2">
-          <NumberField
-            label="Width"
-            value={primitive.type === "text" ? bounds.width : primitive.width ?? bounds.width}
-            onChange={(value) => onChange("width", value)}
-          />
-          <NumberField
-            label="Height"
-            value={primitive.height ?? bounds.height}
-            onChange={(value) => onChange("height", value)}
-          />
-        </div>
-      ) : null}
-
-      {primitive.type === "text" ? (
-        <>
-          <TextField
-            label="Text"
-            value={primitive.text || ""}
-            onChange={(value) => {
-              onChange("text", value);
-              if (value) {
-                onChange("binding", "");
-              }
-            }}
-          />
-          <SelectField
-            label="Binding"
-            value={primitive.binding || ""}
-            onChange={(value) => {
-              onChange("binding", value);
-              if (value) {
-                onChange("text", "");
-              }
-            }}
-            options={[
-              ["", "None"],
-              ["label", "Label"],
-              ["session", "Session"],
-              ["weekly", "Weekly"],
-              ["reset", "Reset"],
-              ["usageMode", "Mode"],
-              ["time", "Time"],
-              ["date", "Date"],
-            ]}
-          />
-          <div className="grid grid-cols-2 gap-2">
-            <NumberField
-              label="Font size"
-              value={primitive.fontSize ?? 2}
-              onChange={(value) => onChange("fontSize", value)}
-            />
-            <SelectField
-              label="Align"
-              value={primitive.align || "left"}
-              onChange={(value) => onChange("align", value)}
-              options={[
-                ["left", "Left"],
-                ["center", "Center"],
-                ["right", "Right"],
-              ]}
-            />
-          </div>
-          <ColorField
-            label="Text color"
-            value={primitive.color || "#FFFFFF"}
-            onChange={(value) => onChange("color", value)}
-          />
-          <div className="grid gap-2">
-            <span className="text-xs font-black uppercase tracking-normal text-[#444933]">
-              Variables
-            </span>
-            <div className="grid grid-cols-2 gap-2">
-              {VARIABLE_TOKENS.map((item) => (
-                <button
-                  className="min-w-0 border border-[#747A60] bg-[#F9F9F9] px-2 py-2 text-left text-xs text-[#1B1B1B] outline-none transition hover:bg-[#EEEEEE] focus-visible:border-[#5E7200]"
-                  key={item.token}
-                  onClick={() => onInsertToken(item.token)}
-                  type="button"
-                >
-                  <span className="block truncate font-black">{item.label}</span>
-                  <code className="block truncate text-[11px] text-[#5E7200]">
-                    {item.token}
-                  </code>
-                </button>
-              ))}
-            </div>
-          </div>
-        </>
-      ) : null}
-
-      {primitive.type === "progress" ? (
-        <>
-          <SelectField
-            label="Binding"
-            value={primitive.binding || "session"}
-            onChange={(value) => onChange("binding", value)}
-            options={[
-              ["session", "Session"],
-              ["weekly", "Weekly"],
-            ]}
-          />
-          <SelectField
-            label="Style"
-            value={primitive.progressStyle || "solid"}
-            onChange={(value) =>
-              onChange("progressStyle", value === "solid" ? "" : value)
-            }
-            options={[
-              ["solid", "Solid"],
-              ["segments", "Segments"],
-            ]}
-          />
-          {primitive.progressStyle === "segments" ? (
-            <div className="grid grid-cols-2 gap-2">
-              <NumberField
-                label="Segments"
-                value={primitive.segments ?? 12}
-                onChange={(value) => onChange("segments", value)}
-              />
-              <NumberField
-                label="Gap"
-                value={primitive.segmentGap ?? 1}
-                onChange={(value) => onChange("segmentGap", value)}
-              />
-            </div>
-          ) : null}
-          <ColorField
-            label="Bar color"
-            value={primitive.color || "#C7FF68"}
-            onChange={(value) => onChange("color", value)}
-          />
-          <ColorField
-            label="Track color"
-            value={primitive.bgColor || "#111111"}
-            onChange={(value) => onChange("bgColor", value)}
-          />
-          <ColorField
-            label="Border color"
-            value={primitive.borderColor || "#3B4552"}
-            onChange={(value) => onChange("borderColor", value)}
-          />
-        </>
-      ) : null}
-
-      {primitive.type === "rect" ? (
-        <ColorField
-          label="Fill color"
-          value={primitive.color || "#222222"}
-          onChange={(value) => onChange("color", value)}
-        />
-      ) : null}
-
-      {primitive.type === "gif" || primitive.type === "sprite" ? (
-        <TextField
-          label="Asset path"
-          value={primitive.assetPath || ""}
-          onChange={(value) => onChange("assetPath", value)}
-        />
-      ) : null}
-
-      {primitive.type === "sprite" ? (
-        <div className="grid grid-cols-3 gap-2">
-          <NumberField
-            label="Frames"
-            value={primitive.frameCount ?? 1}
-            onChange={(value) => onChange("frameCount", value)}
-          />
-          <NumberField
-            label="FPS"
-            value={primitive.fps ?? DEFAULT_SPRITE_FPS}
-            onChange={(value) => onChange("fps", value)}
-          />
-          <NumberField
-            label="Columns"
-            value={primitive.sheetColumns ?? primitive.frameCount ?? 1}
-            onChange={(value) => onChange("sheetColumns", value)}
-          />
-        </div>
-      ) : null}
-
-      <button
-        className="mt-1 inline-flex min-h-11 items-center justify-center gap-2 border border-[#7D2633] bg-[#FFE3E8] px-3 text-sm font-black text-[#7D2633] outline-none hover:bg-[#FFD1DA] focus-visible:border-[#7D2633]"
-        onClick={onDelete}
-        type="button"
-      >
-        <Trash2 size={16} aria-hidden />
-        <span>Delete</span>
-      </button>
-    </div>
-  );
-}
-
-function DarkButton({
-  busy = false,
-  className = "",
-  disabled = false,
-  fullWidth = true,
-  icon,
-  label,
-  onClick,
-  variant = "default",
-}: {
-  busy?: boolean;
-  className?: string;
-  disabled?: boolean;
-  fullWidth?: boolean;
-  icon: ReactNode;
-  label: string;
-  onClick: () => void;
-  variant?: "accent" | "default";
-}) {
-  return (
-    <ControlCenterButton
-      busy={busy}
-      className={className}
-      disabled={disabled}
-      fullWidth={fullWidth}
-      icon={icon}
-      label={label}
-      onClick={onClick}
-      size="default"
-      variant={variant === "accent" ? "primary" : "secondary"}
-    />
-  );
-}
-
-function AddButton({
-  icon,
-  label,
-  onClick,
-}: {
-  icon: ReactNode;
-  label: string;
-  onClick: () => void;
-}) {
-  return (
-    <button
-      className="grid min-h-14 min-w-0 place-items-center gap-1 border border-[#747A60] bg-[#F9F9F9] p-1.5 text-center text-[10px] font-black text-[#1B1B1B] outline-none hover:bg-[#EEEEEE] focus-visible:border-[#5E7200]"
-      onClick={onClick}
-      type="button"
-    >
-      <span className="grid size-7 place-items-center border border-[#747A60] text-[#5E7200]">
-        {icon}
-      </span>
-      <span className="max-w-full truncate">{label}</span>
-    </button>
-  );
-}
-
-function AssetRow({
-  asset,
-  onRemove,
-  onUse,
-  path,
-  referenced,
-}: {
-  asset: ThemeStudioAsset;
-  onRemove: () => void;
-  onUse: () => void;
-  path: string;
-  referenced: boolean;
-}) {
-  return (
-    <div className="grid gap-2 border border-[#747A60] bg-[#F9F9F9] p-2">
-      <div className="min-w-0">
-        <div className="truncate text-xs font-black text-[#1B1B1B]">
-          {assetFileName(path)}
-        </div>
-        <div className="mt-1 flex min-w-0 flex-wrap gap-2 text-[11px] text-[#444933]">
-          <span>{assetKindLabel(path)}</span>
-          <span>{formatBytes(themeAssetByteLength(asset))}</span>
-          <span className={referenced ? "text-[#5E7200]" : "text-[#8A6D00]"}>
-            {referenced ? "Used" : "Unused"}
-          </span>
-        </div>
-        <code className="mt-1 block truncate text-[11px] text-[#5E7200]">
-          {path}
-        </code>
-      </div>
-      <div className="grid grid-cols-2 gap-2">
-        <button
-          className="min-h-10 border border-[#5E7200] bg-[#F1FFD0] px-2 text-xs font-black text-[#3B5200] outline-none hover:bg-[#E3FFA6] focus-visible:border-[#5E7200]"
-          onClick={onUse}
-          type="button"
-        >
-          Use
-        </button>
-        <button
-          className="min-h-10 border border-[#7D2633] bg-[#FFE3E8] px-2 text-xs font-black text-[#7D2633] outline-none hover:bg-[#FFD1DA] focus-visible:border-[#7D2633]"
-          onClick={onRemove}
-          type="button"
-        >
-          Remove
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function PanelTitle({ icon, title }: { icon: ReactNode; title: string }) {
-  return (
-    <div className="mb-3 flex items-center gap-2 text-xs font-black uppercase tracking-normal text-[#1B1B1B]">
-      <span className="text-[#5E7200]">{icon}</span>
-      <span>{title}</span>
-    </div>
-  );
-}
-
 export function clearRetiredAiThemeStorage() {
   try {
     const keys: string[] = [];
@@ -2265,345 +1500,4 @@ export function clearRetiredAiThemeStorage() {
   } catch {
     // Storage can be unavailable in hardened browser modes. The AI feature stays absent.
   }
-}
-
-function prettyJson(spec: ThemeStudioSpec): string {
-  return JSON.stringify(JSON.parse(deviceThemeSpecJson(spec)), null, 2);
-}
-
-function defaultPrimitive(
-  type: "progress" | "rect" | "text",
-  count: number,
-): ThemeStudioPrimitive {
-  const offset = Math.min(42, count * 8);
-  if (type === "text") {
-    return {
-      type,
-      x: 24 + offset,
-      y: 24 + offset,
-      text: "Text",
-      fontSize: 2,
-      color: "#EEF2F6",
-    };
-  }
-  if (type === "progress") {
-    return {
-      type,
-      x: 24,
-      y: 118 + Math.min(40, offset),
-      width: 190,
-      height: 16,
-      binding: "session",
-      color: "#C7FF68",
-      bgColor: "#111111",
-      borderColor: "#3B4552",
-    };
-  }
-  return {
-    type,
-    x: 32 + offset,
-    y: 32 + offset,
-    width: 56,
-    height: 36,
-    color: "#24313D",
-  };
-}
-
-function setPrimitiveField(
-  primitive: ThemeStudioPrimitive,
-  field: FieldKey,
-  value: unknown,
-) {
-  if (value === "") {
-    delete primitive[field];
-    return;
-  }
-  (primitive as Record<FieldKey, unknown>)[field] = value;
-}
-
-function primitiveBounds(primitive: ThemeStudioPrimitive) {
-  if (
-    primitive.type === "rect" ||
-    primitive.type === "progress" ||
-    primitive.type === "gif" ||
-    primitive.type === "sprite" ||
-    primitive.type === "pixels"
-  ) {
-    return {
-      height: primitive.height || 16,
-      width: primitive.width || 16,
-    };
-  }
-  const fontSize = primitive.fontSize || 1;
-  const visibleWidth = Math.max(
-    primitive.width || 0,
-    textPrimitiveNaturalWidth(primitive, fontSize),
-  );
-  const visibleHeight = textPrimitiveSelectionHeight(primitive, fontSize);
-  return {
-    height: Math.min(
-      Math.max(1, DISPLAY_SIZE - primitive.y),
-      Math.max(8, visibleHeight),
-    ),
-    width: Math.min(
-      Math.max(1, DISPLAY_SIZE - primitive.x),
-      Math.max(24, visibleWidth),
-    ),
-  };
-}
-
-function aspectLockedResizeSize({
-  maxHeight,
-  maxWidth,
-  originHeight,
-  originWidth,
-  targetHeight,
-  targetWidth,
-}: {
-  maxHeight: number;
-  maxWidth: number;
-  originHeight: number;
-  originWidth: number;
-  targetHeight: number;
-  targetWidth: number;
-}): ResizeSize {
-  if (originWidth <= 0 || originHeight <= 0) {
-    return {
-      height: clampInt(targetHeight, 1, maxHeight),
-      width: clampInt(targetWidth, 1, maxWidth),
-    };
-  }
-  const widthScale = targetWidth / originWidth;
-  const heightScale = targetHeight / originHeight;
-  const dominantScale =
-    Math.abs(widthScale - 1) >= Math.abs(heightScale - 1)
-      ? widthScale
-      : heightScale;
-  const minScale = Math.max(1 / originWidth, 1 / originHeight);
-  const maxScale = Math.min(maxWidth / originWidth, maxHeight / originHeight);
-  const scale = Math.max(minScale, Math.min(maxScale, dominantScale));
-
-  return {
-    height: clampInt(originHeight * scale, 1, maxHeight),
-    width: clampInt(originWidth * scale, 1, maxWidth),
-  };
-}
-
-function clampedMoveDelta(
-  origins: DragMoveOrigin[],
-  rawDeltaX: number,
-  rawDeltaY: number,
-) {
-  const minDeltaX = Math.max(...origins.map((origin) => -origin.x));
-  const maxDeltaX = Math.min(
-    ...origins.map((origin) => DISPLAY_SIZE - origin.x - origin.width),
-  );
-  const minDeltaY = Math.max(...origins.map((origin) => -origin.y));
-  const maxDeltaY = Math.min(
-    ...origins.map((origin) => DISPLAY_SIZE - origin.y - origin.height),
-  );
-
-  return {
-    x: clampInt(rawDeltaX, minDeltaX, maxDeltaX),
-    y: clampInt(rawDeltaY, minDeltaY, maxDeltaY),
-  };
-}
-
-function normalizeSelectedIndices(indices: number[], primitiveCount: number): number[] {
-  return [...new Set(indices)].filter(
-    (index) => Number.isInteger(index) && index >= 0 && index < primitiveCount,
-  );
-}
-
-function normalizedSelectionBox(
-  startX: number,
-  startY: number,
-  currentX: number,
-  currentY: number,
-): SelectionBox {
-  const minX = Math.max(0, Math.min(startX, currentX));
-  const minY = Math.max(0, Math.min(startY, currentY));
-  const maxX = Math.min(DISPLAY_SIZE, Math.max(startX, currentX));
-  const maxY = Math.min(DISPLAY_SIZE, Math.max(startY, currentY));
-  return {
-    height: Math.max(0, maxY - minY),
-    width: Math.max(0, maxX - minX),
-    x: minX,
-    y: minY,
-  };
-}
-
-function selectedPrimitiveIndices(
-  primitives: ThemeStudioPrimitive[],
-  selection: SelectionBox,
-): number[] {
-  const hits = primitives.flatMap((primitive, index) => {
-    const bounds = primitiveBounds(primitive);
-    const box = {
-      height: bounds.height,
-      width: bounds.width,
-      x: primitive.x,
-      y: primitive.y,
-    };
-    return selectionBoxesIntersect(selection, box) ? [index] : [];
-  });
-  const foregroundHits = hits.filter((index) => !isFullCanvasRect(primitives[index]));
-  return foregroundHits.length > 0 ? foregroundHits : hits;
-}
-
-function selectionBoxesIntersect(a: SelectionBox, b: SelectionBox): boolean {
-  return (
-    a.x < b.x + b.width &&
-    a.x + a.width > b.x &&
-    a.y < b.y + b.height &&
-    a.y + a.height > b.y
-  );
-}
-
-function isFullCanvasRect(primitive: ThemeStudioPrimitive | undefined): boolean {
-  return Boolean(
-    primitive &&
-      primitive.type === "rect" &&
-      primitive.x === 0 &&
-      primitive.y === 0 &&
-      primitive.width === DISPLAY_SIZE &&
-      primitive.height === DISPLAY_SIZE,
-  );
-}
-
-function isEditableKeyboardTarget(target: EventTarget | null): boolean {
-  if (!(target instanceof HTMLElement)) {
-    return false;
-  }
-  const tagName = target.tagName.toLowerCase();
-  return (
-    target.isContentEditable ||
-    tagName === "a" ||
-    tagName === "button" ||
-    tagName === "input" ||
-    tagName === "textarea" ||
-    tagName === "select"
-  );
-}
-
-function textPrimitiveNaturalWidth(
-  primitive: ThemeStudioPrimitive,
-  fontSize = primitive.fontSize || 1,
-) {
-  const text = primitive.binding
-    ? boundText(primitive.binding)
-    : substituteText(primitive.text || "Text");
-  const renderFontSize = textPrimitiveRenderFontSize(primitive, fontSize);
-  return Math.min(
-    Math.max(1, DISPLAY_SIZE - primitive.x),
-    Math.ceil(text.length * renderFontSize * 0.6),
-  );
-}
-
-function textPrimitiveRenderFontSize(
-  primitive: ThemeStudioPrimitive,
-  fontSize = primitive.fontSize || 1,
-) {
-  const scale = Math.max(1, fontSize);
-  switch (primitive.font || 1) {
-    case 2:
-      return 16 * scale;
-    case 4:
-      return 26 * scale;
-    case 6:
-    case 7:
-      return 48 * scale;
-    case 8:
-      return 75 * scale;
-    case 1:
-    default:
-      return 8 * scale;
-  }
-}
-
-function textPrimitiveSelectionHeight(
-  primitive: ThemeStudioPrimitive,
-  fontSize = primitive.fontSize || 1,
-) {
-  return Math.ceil(
-    textPrimitiveRenderFontSize(primitive, fontSize) *
-      TEXT_SELECTION_HEIGHT_SCALE,
-  );
-}
-
-function textPrimitiveFontSizeFromVisualHeight(
-  primitive: ThemeStudioPrimitive,
-  height: number,
-) {
-  const baseHeight = textPrimitiveRenderFontSize({ ...primitive, fontSize: 1 }, 1);
-  return Math.round(height / (baseHeight * TEXT_SELECTION_HEIGHT_SCALE));
-}
-
-function primitiveTitle(primitive: ThemeStudioPrimitive): string {
-  if (primitive.type === "text") {
-    return primitive.text || primitive.binding || "Text";
-  }
-  if (primitive.type === "progress") {
-    return primitive.binding || "session";
-  }
-  if (primitive.type === "gif" || primitive.type === "sprite") {
-    return primitive.assetPath?.split("/").pop() || "Asset";
-  }
-  if (primitive.type === "pixels") {
-    return `${primitive.width ?? 0}x${primitive.height ?? 0}`;
-  }
-  return primitive.color || "Rect";
-}
-
-function boundText(binding: string): string {
-  switch (binding) {
-    case "label":
-      return DEFAULT_FRAME.label;
-    case "provider":
-      return DEFAULT_FRAME.provider;
-    case "session":
-    case "sessionPercent":
-      return String(DEFAULT_FRAME.session);
-    case "weekly":
-    case "weeklyPercent":
-      return String(DEFAULT_FRAME.weekly);
-    case "reset":
-    case "resetCountdown":
-      return "1h";
-    case "usageMode":
-      return DEFAULT_FRAME.usageMode;
-    case "activity":
-      return DEFAULT_FRAME.activity;
-    case "time":
-      return DEFAULT_FRAME.time;
-    case "date":
-      return DEFAULT_FRAME.date;
-    case "sessionTokens":
-      return String(DEFAULT_FRAME.sessionTokens);
-    case "weekTokens":
-      return String(DEFAULT_FRAME.weekTokens);
-    case "totalTokens":
-      return String(DEFAULT_FRAME.totalTokens);
-    default:
-      return "";
-  }
-}
-
-function substituteText(value: string): string {
-  return value.replace(/\{([a-zA-Z0-9_-]+)\}/g, (_match, key: string) =>
-    boundText(key),
-  );
-}
-
-function clampInt(value: number, min: number, max: number): number {
-  const rounded = Math.round(value);
-  return Math.max(min, Math.min(Math.max(min, max), rounded));
-}
-
-function titleFromThemeId(themeId: string): string {
-  return themeId
-    .split(/[-_]/)
-    .filter(Boolean)
-    .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
-    .join(" ");
 }
